Green Software Development
Building Sustainable Tools and Frameworks for Reducing Carbon Footprint in Software Development. Building intelligent carbon calculators in code and software that can help measure the carbon footprint of software and provide actionable insights to reduce carbon emissions. The goal is to create a sustainable software development ecosystem that minimizes the impact of technology on the environment.